requirements.txt
psycopg2>=2.8.6,<2.9
pillow>=8.2.0,<8.3.0

asgiref==3.4.1
Django>=3.2.4, <3.3
pytz==2021.3
sqlparse==0.4.2
graphene-django==2.15.0

django-graphql-jwt==0.3.0
PyJWT==1.7.0
graphene==2.1.8
django-filter
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3

